Introduce authentication with JWT and Cookie
The current goal is to ensure all access to the API are authenticated using a secure JWT which can be generated from the
command-line using a secrets file so that individual users of sensei can host a running instance on a remote machine securely.
Tests are passing using a default valid token but we need to wire in the configuration from the command-line.