[wip] Start adding uuid to identify events
Ran into the problem of returning an index in the EventView that's the raw index of the inserted event, which is wrong because this
does not account for different users: I store the events for all users in the same table.
This is great as the Model actually catches some inconsistency in the underlying implementation. I suspect the only right way to
solve this is to have one table per user? Or perhaps, storing an index of last event inserted per user would be enough?