Experiment with C TLS client
The idea is to build a client that can be easily and quickly deployed
on various machines to be used to provide data to sensei, so a small C
program that can be compiled without too much hassle on various
platforms would be ideal.

Here I only wrote a SSL client that can successfully connect to a
HTTPS server which is not super helpful as HTTP protocol is complex,
requires JSON, etc.

So the idea would be to have sensei expose a TCP port (behind TLS)
that clients could connect to (using same authentication mechanism as
the web server), using a "simpler" binary interface that would make
life easier for C.