Remove one layer of indirection for handling events in Backends
Instead of mapping backned types to a specialised `BackendIO` that
only handles `ClientMonad` requests, the top-level app maps backends
to a `BackendHandler` that directly handles the `Event` in a monadic
context. This merges the `IsBackend` class and `BackendIO` structure
into a single function.

Then one needs to pass a function to properly handle the ClientMonad
for the Bsky client to construct a `BackendHandler` for Bsky. This
will be tested in a different test.