Do not propagate event to backends if DB write fails
Writing this test required a whole machinery to provide a fake
implementation of DB interface, plus restructuring the App to be able
to inject any DB instance, but I think it's worth it because it
enables testing the whole application in presence of arbitrary
failures of arbitrary operations.

This is something I should have done earlier, or at least designed the
application for by introducing those db failures test earlier but it
is what it is.