Lookup explicitly for backend handler in IO monad
This fixes the problem where the server cannot find the backend even
though it's configured because it's created within IO monad, in the
'sensei' function. It should be initialized lower in the stack, in the
AppM monad where all functions in 'Server' live but unfortunately this
is not trivial to fix, so for the time being I will resort to this
kludge.